libutil: move utility functions to libutil
Problem: it's not entirely obvious what parts
of the project are integral parts vs standalone,
separately testable and reusable modules.

Move parts from the latter category to a libutil
directory.
