+++ src/getquota.c
@@ -35,7 +35,8 @@
 #include <pwd.h>
 #include <assert.h>
 
-#include "util.h"
+#include "src/libutil/util.h"
+
 #include "getquota.h"
 #include "getquota_private.h"
