+++ src/libutil/Makefile.am
@@ -1,0 +1,14 @@
+AM_CFLAGS = @GCCWARN@
+
+AM_CPPFLAGS = \
+	-I$(top_srcdir)
+
+noinst_LIBRARIES = libutil.a
+
+libutil_a_SOURCES = \
+	listint.c \
+	listint.h \
+	util.c \
+	util.h \
+	getconf.c \
+	getconf.h
