+++ src/libutil/getconf.c
