+++ src/libutil/getconf.h
