+++ src/libutil/listint.c
