+++ src/libutil/util.c
