+++ src/libutil/util.h
