+++ src/repquota.c
@@ -43,10 +43,11 @@
 #include <libgen.h>
 #include <sys/stat.h>
 
-#include "getconf.h"
+#include "src/libutil/getconf.h"
+#include "src/libutil/util.h"
+#include "src/libutil/listint.h"
+
 #include "getquota.h"
-#include "listint.h"
-#include "util.h"
 
 static void usage(void);
 static void add_quota(confent_t *cp, List qlist, uid_t uid, char *name);
